Move declare v2 test to separate file [skip ci]
